♻️ remove interface + split error codes
